# openapi.yaml
openapi: 3.1.0
info:
  title: OKR Tracking System API
  description: |
    ## OKR Tracking System — Garant Bank

    REST API for managing **Objectives & Key Results (OKR)**, organizational structure, user management, and multi-source performance evaluations.

    ### Key Features
    - **OKR Management** — Create and track Objectives with weighted Key Results
    - **Multi-Source Evaluations** — Director (star rating), HR (letter grade), and Business Block assessments
    - **Score Calculation** — Automatic weighted scoring: 60% OKR + 20% Director + 20% HR
    - **Organizational Hierarchy** — Divisions → Departments → Objectives → Key Results
    - **Role-Based Access Control** — 6 roles with fine-grained permissions
    - **Excel Export** — Export all OKR data to Excel spreadsheets

    ### Authentication
    All endpoints (except login and export) require a **JWT Bearer token**.
    1. Call `POST /api/auth/login` with your credentials
    2. Copy the `token` from the response
    3. Click the **Authorize** button above and enter the token

    ### Roles
    | Role | Description |
    |------|-------------|
    | `ADMIN` | Full system access, user management |
    | `DIRECTOR` | Division oversight, star-based evaluations |
    | `DEPARTMENT_LEADER` | Department OKR management |
    | `HR` | Letter-grade evaluations (A/B/C/D) |
    | `BUSINESS_BLOCK` | Business block star-based evaluations |
    | `EMPLOYEE` | View own profile and assigned departments |
  version: 1.0.0
...
servers:
...
security:
- Bearer Authentication: []
...
paths:
...
  /api/evaluations/{id}:
    put:
      tags:
      - Evaluations
      summary: Update evaluation
      description: Modify an existing evaluation. Evaluators can update their evaluations
        even after submission (to correct mistakes). Only the original evaluator can
        update.
      operationId: updateEvaluation
      parameters:
      - name: id
        in: path
        description: Evaluation ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationCreateRequest"
        required: true
      responses:
        "200":
          description: Evaluation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationDTO"
        "403":
          description: Not the evaluation owner
        "404":
          description: Evaluation not found
    delete:
      tags:
      - Evaluations
      summary: Delete evaluation
      description: Delete a draft evaluation. Only evaluations in `DRAFT` status can
        be deleted. Only the original evaluator can delete their evaluation.
      operationId: deleteEvaluation
      parameters:
      - name: id
        in: path
        description: Evaluation ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Evaluation deleted
        "400":
          description: Cannot delete non-draft evaluation
        "403":
          description: Not the evaluation owner
        "404":
          description: Evaluation not found
...
  /api/evaluations:
    post:
      tags:
      - Evaluations
      summary: Create evaluation
      description: |-
        Create a new performance evaluation for a department or employee.

        **Evaluator types and their rating scales:**
        - `DIRECTOR` — 1-5 star rating (converted to 4.25-5.0 score)
        - `HR` — Letter grade: A (5.0), B (4.75), C (4.5), D (4.25)
        - `BUSINESS_BLOCK` — 1-5 star rating (displayed separately, not in final calculation)

        Evaluations are created in `DRAFT` status and must be submitted separately.

        **Requires:** DIRECTOR, HR, BUSINESS_BLOCK, or ADMIN role.
      operationId: createEvaluation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationCreateRequest"
        required: true
      responses:
        "200":
          description: Evaluation created in DRAFT status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationDTO"
        "400":
          description: "Invalid request (e.g., invalid rating)"
        "403":
          description: Insufficient permissions
  /api/evaluations/{id}/submit:
    post:
      tags:
      - Evaluations
      summary: Submit evaluation
      description: "Change an evaluation from `DRAFT` to `SUBMITTED` status. Once\
        \ submitted, the evaluation contributes to the department/employee's final\
        \ score calculation. Only the original evaluator can submit their own evaluation."
      operationId: submitEvaluation
      parameters:
      - name: id
        in: path
        description: Evaluation ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Evaluation submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationDTO"
        "403":
          description: Not the evaluation owner
        "404":
          description: Evaluation not found
...
  /api/evaluations/target/{type}/{id}:
    get:
      tags:
      - Evaluations
      summary: Get evaluations for target
      description: "Retrieve all evaluations for a specific department or employee.\
        \ Returns evaluations from all evaluator types (Director, HR, Business Block)."
      operationId: getEvaluationsForTarget
      parameters:
      - name: type
        in: path
        description: Target type
        required: true
        schema:
          type: string
          enum:
          - DEPARTMENT
          - EMPLOYEE
      - name: id
        in: path
        description: Target ID (UUID)
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: List of evaluations for the target
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvaluationDTO"
...
  /api/evaluations/all:
    get:
      tags:
      - Evaluations
      summary: Get all evaluations
      description: Retrieve all evaluations in the system. **Requires ADMIN role.**
        Intended for debugging and administration.
      operationId: getAllEvaluations
      responses:
        "200":
          description: All evaluations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvaluationDTO"
        "403":
          description: Only ADMIN can view all evaluations
...
components:
  schemas:
...
    EvaluationCreateRequest:
      type: object
      description: Request to create or update an evaluation
      properties:
        targetType:
          type: string
          description: Target entity type
          enum:
          - DEPARTMENT
          - EMPLOYEE
          example: DEPARTMENT
        targetId:
          type: string
          format: uuid
          description: ID of the target department or employee
          example: 550e8400-e29b-41d4-a716-446655440000
        evaluatorType:
          type: string
          description: Evaluator type
          enum:
          - DIRECTOR
          - HR
          - BUSINESS_BLOCK
          example: DIRECTOR
        numericRating:
          type: number
          format: double
          description: Numeric rating — for DIRECTOR (4.25-5.0) or BUSINESS_BLOCK
            (4.25-5.0)
          example: 4.75
        starRating:
          type: integer
          format: int32
          description: "Star rating — for DIRECTOR and BUSINESS_BLOCK (1-5 stars,\
            \ converted to 4.25-5.0)"
          example: 4
        letterRating:
          type: string
          description: Letter rating — for HR evaluations
          enum:
          - A
          - B
          - C
          - D
          example: B
        comment:
          type: string
          description: Evaluator's comment
          example: Good progress on quarterly objectives
      required:
      - evaluatorType
      - targetId
      - targetType
    EvaluationDTO:
      type: object
      description: "Evaluation record from a Director, HR, or Business Block evaluator"
      properties:
        id:
          type: string
          format: uuid
          description: Evaluation ID
          example: 550e8400-e29b-41d4-a716-446655440000
        evaluatorId:
          type: string
          format: uuid
          description: ID of the user who created this evaluation
          example: 660e8400-e29b-41d4-a716-446655440001
        evaluatorName:
          type: string
          description: Name of the evaluator
          example: Jane Smith
        evaluatorType:
          type: string
          description: Type of evaluator
          enum:
          - DIRECTOR
          - HR
          - BUSINESS_BLOCK
          example: DIRECTOR
        targetType:
          type: string
          description: Target entity type
          enum:
          - DEPARTMENT
          - EMPLOYEE
          example: DEPARTMENT
        targetId:
          type: string
          format: uuid
          description: ID of the target department or employee
          example: 770e8400-e29b-41d4-a716-446655440002
        numericRating:
          type: number
          format: double
          description: "Numeric rating (Director: 4.25-5.0 from stars; HR: 4.25-5.0\
            \ from letter grade)"
          example: 4.75
        letterRating:
          type: string
          description: Letter rating for HR evaluations
          enum:
          - A
          - B
          - C
          - D
          example: B
        comment:
          type: string
          description: Evaluator's comment
          example: "Strong performance, exceeds expectations in key areas"
        status:
          type: string
          description: Evaluation status
          enum:
          - DRAFT
          - SUBMITTED
          - APPROVED
          example: SUBMITTED
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        submittedAt:
          type: string
          format: date-time
          description: Submission timestamp (null if still in DRAFT)
...
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Enter the JWT token obtained from `POST /api/auth/login`
      name: Bearer Authentication
      scheme: bearer
      bearerFormat: JWT